Browser choice when opening links in About view.
